chore(presentation): move defs
